+++ src/Transport/Connection/Heartbeat.php
@@ -9,7 +9,7 @@ use Innmind\AMQP\{
 use Innmind\TimeContinuum\{
     Clock,
     PointInTime,
-    ElapsedPeriod,
+    Period,
 };
 use Innmind\Immutable\Sequence;
 
@@ -19,12 +19,12 @@ use Innmind\Immutable\Sequence;
 final class Heartbeat
 {
     private Clock $clock;
-    private ElapsedPeriod $threshold;
+    private Period $threshold;
     private PointInTime $lastReceivedData;
 
     private function __construct(
         Clock $clock,
-        ElapsedPeriod $threshold,
+        Period $threshold,
         PointInTime $lastReceivedData,
     ) {
         $this->clock = $clock;
@@ -32,7 +32,7 @@ final class Heartbeat
         $this->lastReceivedData = $lastReceivedData;
     }
 
-    public static function start(Clock $clock, ElapsedPeriod $threshold): self
+    public static function start(Clock $clock, Period $threshold): self
     {
         return new self($clock, $threshold, $clock->now());
     }
@@ -47,7 +47,7 @@ final class Heartbeat
                 ->clock
                 ->now()
                 ->elapsedSince($this->lastReceivedData)
-                ->longerThan($this->threshold)
+                ->longerThan($this->threshold->asElapsedPeriod())
         ) {
             $this->lastReceivedData = $this->clock->now();
 
@@ -62,7 +62,7 @@ final class Heartbeat
         $this->lastReceivedData = $this->clock->now();
     }
 
-    public function adjust(ElapsedPeriod $threshold): self
+    public function adjust(Period $threshold): self
     {
         return new self(
             $this->clock,