+++ tests/Transport/ProtocolTest.php
@@ -82,7 +82,7 @@ class ProtocolTest extends TestCase
                         ->withPriority(Priority::five)
                         ->withCorrelationId(CorrelationId::of('correlation'))
                         ->withReplyTo(ReplyTo::of('reply'))
-                        ->withExpiration(Period::second(1)->asElapsedPeriod())
+                        ->withExpiration(Period::second(1))
                         ->withId(Id::of('id'))
                         ->withTimestamp($now = PointInTime::now())
                         ->withType(Type::of('type'))