+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changed
 
 - Requires `innmind/foundation:~1.5`
+- `Innmind\AMQP\Factory::make()` timeout argument is now expressed via `Innmind\TimeContinuum\Period`
+- `Innmind\AMQP\Model\Basic\Message` expiration is now expressed via `Innmind\TimeContinuum\Period`
 
 ### Fixed